separate mod_pow and main
